django_test/stackover_test/requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.7
django-autocomplete-light==3.8.2
django-countries==7.1
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
django-humanize==0.1.2
django-tables2==2.3.4
django-taggit==1.3.0
django-tailwind==1.2.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7.1
humanize==3.2.0
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
numpy==1.20.1
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
